teamkerbell_backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-core==1.4.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
uritemplate==4.1.1
